refactor(telemetry): collapse org tier and drop shared reason module
